+++ src/tests/test_dot.cc
@@ -3,20 +3,20 @@
 #include <boost/test/included/unit_test.hpp>
 
 BOOST_AUTO_TEST_CASE(dot_product) {
-  eigencuda::EigenCuda EC;
+  eigencuda::CudaPipeline CP;
   Eigen::MatrixXd A = Eigen::MatrixXd::Zero(2, 2);
   Eigen::MatrixXd B = Eigen::MatrixXd::Zero(2, 2);
 
   A << 1., 2., 3., 4.;
   B << 5., 6., 7., 8.;
 
-  Eigen::MatrixXd C = EC.matrix_mult(A, B);
+  Eigen::MatrixXd C = CP.matrix_mult(A, B);
   BOOST_CHECK_EQUAL(C.sum(), 134.);
 }
 
 BOOST_AUTO_TEST_CASE(right_matrix_multiplication) {
   // Call the class to handle GPU resources
-  eigencuda::EigenCuda EC;
+  eigencuda::CudaPipeline CP;
 
   // Call matrix multiplication GPU
   Eigen::MatrixXd A = Eigen::MatrixXd::Zero(2, 2);
@@ -37,7 +37,7 @@ BOOST_AUTO_TEST_CASE(right_matrix_multiplication) {
   Z << 55., 82., 63., 94., 71., 106.;
 
   std::vector<Eigen::MatrixXd> tensor{B, C, D};
-  EC.right_matrix_tensor_mult(std::move(tensor), A);
+  CP.right_matrix_tensor_mult(tensor, A);
 
   // Expected results
   BOOST_TEST(X.isApprox(tensor[0]));
@@ -49,16 +49,9 @@ BOOST_AUTO_TEST_CASE(wrong_shape_cublas) {
   Eigen::MatrixXd A = Eigen::MatrixXd::Random(2, 2);
   Eigen::MatrixXd B = Eigen::MatrixXd::Random(5, 5);
 
-  eigencuda::EigenCuda EC;
+  eigencuda::CudaPipeline CP;
   std::vector<Eigen::MatrixXd> tensor{B};
-  try {
-    EC.right_matrix_tensor_mult(std::move(tensor), A);
-  } catch (const std::runtime_error& error) {
-    std::string error_msg = error.what();
-    std::string reason = "an illegal value";
-    if (error_msg.find(reason) != std::string::npos) {
-      auto eptr = std::current_exception();
-      std::rethrow_exception(eptr);
-    }
-  }
+
+  BOOST_REQUIRE_THROW(CP.right_matrix_tensor_mult(tensor, A),
+                      std::runtime_error);
 }